+++ src/global.d.ts
@@ -1,2 +1,10 @@
 declare const IS_PROD: boolean;
 declare const IS_DEV: boolean;
+
+declare module 'html360-gen' {
+  export interface Html360Gen {
+    getBinaryPath(): string;
+  }
+
+  export const html360Gen: Html360Gen;
+}